style(frontend): redesigned all pages related to roles
First, on the page with all roles, I fixed the add role button. Second, on the role page, I redesigned the permissions button and added a card with information about the role. Thirdly, I redesigned the role creation page to be more similar to the second point. The last step was to add a clearer text to the permissions modal window.