feat(role): added updating role description and name
Added put request to RoleView where update data. Expanded onChange and replaced it with onNameChange and onDescChange for changing text in different fields. Added nameIsChanged and descIsChanged to track changes. Added a new request to onSubmit. To change the description, I made a button that opens a Textarea with automatic focus.